requirements.txt
# World Market Shopping Agent Dependencies

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Claude API
anthropic>=0.39.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# UI (optional - Streamlit for simple interface)
streamlit>=1.31.0

# Data handling
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
pyyaml>=6.0.0

# Per-customer prompt optimization
gepa>=0.1.0
